优先使用int替换为uint，方便运算，能够准确识别产生的负数。 GCC/VC编译通过，MDK编译Port/I2C失败

+++ Device/Proxy.h
@@ -53,9 +53,9 @@ public:
 
 	SerialPort port;
 
-	ushort	parity;
-	ushort	dataBits;
-	ushort	stopBits;
+	byte	parity;
+	byte	dataBits;
+	byte	stopBits;
 	int		baudRate;
 
 	virtual bool OnSetConfig(Dictionary<cstring, int>& config, String& str) override;
